# src/main/java/org/openapitools/api/SolicitudApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.springframework.lang.Nullable;
import org.openapitools.model.ProblemDetails;
import org.openapitools.model.Solicitud;
import org.openapitools.model.SolicitudResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-07T16:11:29.458034+02:00[Europe/Madrid]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "Solicitud", description = "the Solicitud API")
public interface SolicitudApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_SOLICITUD_COMPROBAR_SOLICITUD_GET = "/Solicitud/ComprobarSolicitud";
    /**
     * GET /Solicitud/ComprobarSolicitud
     *
     * @param nombreUsuario  (optional)
     * @param tok  (optional)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "solicitudComprobarSolicitudGet",
        tags = { "Solicitud" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "text/plain", array = @ArraySchema(schema = @Schema(implementation = Integer.class))),
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Integer.class))),
                @Content(mediaType = "text/json", array = @ArraySchema(schema = @Schema(implementation = Integer.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "text/plain", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = ProblemDetails.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SolicitudApi.PATH_SOLICITUD_COMPROBAR_SOLICITUD_GET,
        produces = { "text/plain", "application/json", "text/json" }
    )
    default ResponseEntity<List<Integer>> solicitudComprobarSolicitudGet(
        @Parameter(name = "nombreUsuario", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "nombreUsuario", required = false) @Nullable String nombreUsuario,
        @Parameter(name = "tok", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "tok", required = false) @Nullable Integer tok
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ 0, 0 ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"instance\", \"detail\" : \"detail\", \"type\" : \"type\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/json"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/json";
                    ApiUtil.setExampleResponse(request, "text/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/plain"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/plain";
                    ApiUtil.setExampleResponse(request, "text/plain", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SOLICITUD_GET_SOLICITUDES_USUARIO_GET = "/Solicitud/GetSolicitudesUsuario";
    /**
     * GET /Solicitud/GetSolicitudesUsuario
     *
     * @param nombreUsuario  (optional)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "solicitudGetSolicitudesUsuarioGet",
        tags = { "Solicitud" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "text/plain", array = @ArraySchema(schema = @Schema(implementation = Integer.class))),
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Integer.class))),
                @Content(mediaType = "text/json", array = @ArraySchema(schema = @Schema(implementation = Integer.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "text/plain", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = ProblemDetails.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SolicitudApi.PATH_SOLICITUD_GET_SOLICITUDES_USUARIO_GET,
        produces = { "text/plain", "application/json", "text/json" }
    )
    default ResponseEntity<List<Integer>> solicitudGetSolicitudesUsuarioGet(
        @Parameter(name = "nombreUsuario", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "nombreUsuario", required = false) @Nullable String nombreUsuario
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ 0, 0 ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"instance\", \"detail\" : \"detail\", \"type\" : \"type\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/json"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/json";
                    ApiUtil.setExampleResponse(request, "text/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/plain"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/plain";
                    ApiUtil.setExampleResponse(request, "text/plain", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SOLICITUD_SOLICITAR_POST = "/Solicitud/Solicitar";
    /**
     * POST /Solicitud/Solicitar
     *
     * @param nombreUsuario  (optional)
     * @param solicitud  (optional)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "solicitudSolicitarPost",
        tags = { "Solicitud" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "text/plain", schema = @Schema(implementation = SolicitudResponse.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = SolicitudResponse.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = SolicitudResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "text/plain", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetails.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = ProblemDetails.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = SolicitudApi.PATH_SOLICITUD_SOLICITAR_POST,
        produces = { "text/plain", "application/json", "text/json" },
        consumes = { "application/json", "text/json", "application/*+json" }
    )
    default ResponseEntity<SolicitudResponse> solicitudSolicitarPost(
        @Parameter(name = "nombreUsuario", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "nombreUsuario", required = false) @Nullable String nombreUsuario,
        @Parameter(name = "Solicitud", description = "") @Valid @RequestBody(required = false) @Nullable Solicitud solicitud
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : true, \"tokenSolicitud\" : 0, \"errorMessage\" : \"errorMessage\", \"done\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/json"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/json";
                    ApiUtil.setExampleResponse(request, "text/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/plain"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/plain";
                    ApiUtil.setExampleResponse(request, "text/plain", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"instance\", \"detail\" : \"detail\", \"type\" : \"type\", \"title\" : \"title\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/json"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/json";
                    ApiUtil.setExampleResponse(request, "text/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("text/plain"))) {
                    String exampleString = "Custom MIME type example not yet supported: text/plain";
                    ApiUtil.setExampleResponse(request, "text/plain", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}